# FIX Memory leak in `DefaultRegisteredServicesTracker`