--- a/src/Autofac/Core/Registration/DefaultRegisteredServicesTracker.cs
+++ b/src/Autofac/Core/Registration/DefaultRegisteredServicesTracker.cs
@@ -25,12 +25,12 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
     /// <summary>
     /// External registration sources.
     /// </summary>
-    private readonly List<IRegistrationSource> _dynamicRegistrationSources = new();
+    private readonly Stack<IRegistrationSource> _dynamicRegistrationSources = new();
 
     /// <summary>
     /// All registrations.
     /// </summary>
-    private readonly ConcurrentBag<IComponentRegistration> _registrations = new();
+    private readonly ConcurrentQueue<IComponentRegistration> _registrations = new();
 
     private readonly List<IServiceMiddlewareSource> _servicePipelineSources = new();
 
@@ -71,7 +71,7 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
     {
         get
         {
-            return _registrations.ToList();
+            return _registrations;
         }
     }
 
@@ -80,7 +80,7 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
     {
         get
         {
-            return _dynamicRegistrationSources.ToList();
+            return _dynamicRegistrationSources;
         }
     }
 
@@ -117,7 +117,7 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
             // build pipelines for them.
             // The Registrations collection is only available to consumers once the tracker is contained with a ContainerRegistry
             // and the Complete method has been called.
-            _registrations.Add(registration);
+            _registrations.Enqueue(registration);
             var handler = Registered;
             handler?.Invoke(this, registration);
 
@@ -136,7 +136,7 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
             throw new ArgumentNullException(nameof(source));
         }
 
-        _dynamicRegistrationSources.Insert(0, source);
+        _dynamicRegistrationSources.Push(source);
 
         var handler = RegistrationSourceAdded;
         handler?.Invoke(this, source);
@@ -249,8 +249,6 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
             registration.Dispose();
         }
 
-        ClearRegistrations();
-
         base.Dispose(disposing);
     }
 
@@ -262,24 +260,7 @@ internal class DefaultRegisteredServicesTracker : Disposable, IRegisteredService
             await registration.DisposeAsync().ConfigureAwait(false);
         }
 
-        ClearRegistrations();
-
         // Do not call the base, otherwise the standard Dispose will fire.
-    }
-
-    private void ClearRegistrations()
-    {
-        // If we do not explicitly empty the ConcurrentBag that stores our registrations,
-        // this will cause a memory leak due to threads holding a reference to the bag.
-        // In netstandard2.0 the faster 'Clear' method is not available,
-        // so we have do this manually. We'll use the faster method if it's available though.
-#if NETSTANDARD2_0
-        while (_registrations.TryTake(out _))
-        {
-        }
-#else
-        _registrations.Clear();
-#endif
     }
 
     private ServiceRegistrationInfo GetInitializedServiceInfo(Service service)