- Replaced `ConcurrentBag` with `ConcurrentQueue`. It's causing memory leaks if left undisposed somewhere.
- `ClearRegistrations` is not necessary anymore.
- Decreased allocations on `List.Insert(0,...)` and `.ToList()` calls