@botinko 2024-08-15
In the test environment we are very active in creating containers and child containers. We have found a memory leak. A lot of objects can't be GC'ed because the are active reference from ConcurrentBag<IComponentRegistration>  internal structures (it store data in buckets in thread local storage). So eventually all IComponentRegistration will be gc rooted to static thread local. 

ClearRegistrations wont help much. Just one non-disposed container or child container (which can happens in real world full of bugs) can create a fake gc root for all ConcurrentBag<IComponentRegistration> instances. 

For me, having ConcurrentBag holding IComponentRegistration, that can hold anything in my app is already a performance bug. 

We will make bench for sure, when we finished local testing in our env.