@snaumenko-st 2024-08-15
> Thanks; can you please run our benchmarks for a comparison here, particularly the ones that touch on child unit-of-work scopes with their own dependencies? I'm thinking at least ChildScopeResolveBenchmark and ConcurrencyNestedScopeBenchmark. These sort of collection changes can have unintended perf implications.

Yes, sure thing. Here you are:
```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19045.4651/22H2/2022Update)
11th Gen Intel Core i7-11800H 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.303
  [Host]     : .NET 8.0.7 (8.0.724.31311), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.7 (8.0.724.31311), X64 RyuJIT AVX2


```
### ChildScopeResolveBenchmark

| Method                          | Mean Before | Mean After |
|-------------------------------- |----------:|----------:|
| Resolve                         | 16.944 μs | 16.449  μs |
| ResolveNeverRegisteredFromChild |  7.952 μs |  7.411 μs |

### ConcurrencyBenchmark

| Method                          | ResolveTaskCount | ResolvesPerTask | Mean Before | Mean After |
|-------------------------------- |----------------- |---------------- |---------:|---------:|
| MultipleResolvesOnMultipleTasks | 100              | 100             | 646.6 μs  | 632.3 μs |

### ConcurrencyNestedScopeBenchmark

| Method                          | ConcurrentRequests | RepeatCount | Mean Before | Mean After |
|-------------------------------- |------------------- |------------ |---------:|---------:|
| MultipleResolvesOnMultipleTasks | 100                | 10          | 1.004 ms  | 954.7 μs |

### DeepGraphResolveBenchmark

| Method                         | Mean Before | Mean After |
|-------------------------- |---------------:|-------------:|
| Resolve                         | 7.423 μs         |     7.169 μs |

### RootContainerResolveBenchmark

| Method                         | Mean Before | Mean After |
|-------------------------- |---------------:|-------------:|
| OperatorNew                |   2.131 ns      |   1.990 ns |
| NonSharedReflectionResolve | 208.615 ns | 198.548 ns |
| NonSharedDelegateResolve   | 205.002 ns | 187.762 ns |
| SharedResolve              | 165.256 ns | 159.633 ns |