@snaumenko-st 2024-08-30
This change reverses the order of the Registrations collection unfortunately. However unless somebody depends on that it should be safe.