@alistairjevans 2024-08-30
It was a concurrentbag before right? Which is already technically an unordered set I believe.