@snaumenko-st 2024-08-30
Yes, you are right, but in fact they are selected in reverse order of inserts. So in our case this fact helped to reveal some corner cases of duplicate registrations with different configs, for example. Some unit tests were failing because of that. It's not an issue I believe, but who knows how others are using this interface.